moved various obsolete functions out of main msm_we
